+++ test/unit/internal-indexes.test.ts
@@ -1,0 +1,104 @@
+import assert from 'assert';
+import AsyncTestUtil from 'async-test-util';
+
+import config from './config.ts';
+import * as humansCollection from '../helper/humans-collection.ts';
+import * as schemas from '../helper/schemas.ts';
+import * as schemaObjects from '../helper/schema-objects.ts';
+import {
+    createRxDatabase,
+    randomCouchString,
+    promiseWait,
+    ensureNotFalsy,
+    clone,
+    deepEqual,
+    normalizeMangoQuery,
+    prepareQuery,
+    getChangedDocumentsSinceQuery
+} from '../../plugins/core/index.mjs';
+
+import {
+    first,
+} from 'rxjs/operators';
+import type {
+    RxChangeEvent
+} from '../../plugins/core/index.mjs';
+import { HumanDocumentType } from '../helper/schemas.ts';
+import { firstValueFrom } from 'rxjs';
+
+config.parallel('internal-indexes.test.js', () => {
+
+    async function createCollectionWithInternalIndexes(internalIndexes: string[][], docsAmount: number = 0) {
+        const schema = clone(schemas.human);
+        schema.internalIndexes = internalIndexes;
+        const collection = await humansCollection.createBySchema(
+            schema,
+            'docs'
+        );
+        if (docsAmount > 0) {
+            const docsData = new Array(docsAmount)
+                .fill(0)
+                .map(() => schemaObjects.human());
+            const writeResult = await collection.bulkInsert(docsData);
+            assert.deepStrictEqual(writeResult.error, []);
+        }
+        return collection;
+    }
+
+    describe('creation', () => {
+        it('should allow to use internal indexes and map them correctly', async () => {
+            const myIdx = ['firstName', 'lastName'];
+            const collection = await createCollectionWithInternalIndexes([myIdx]);
+            const foundOnStorage = collection.internalStorageInstance.schema.indexes?.find(idx => deepEqual(idx, myIdx));
+            assert.ok(foundOnStorage);
+            collection.database.destroy();
+        });
+        it('should use the internalIndex in the query planner', async () => {
+            const myIdx = ['firstName', 'lastName'];
+            const collection = await createCollectionWithInternalIndexes([myIdx]);
+
+            const preparedQuery = prepareQuery(
+                collection.internalStorageInstance.schema,
+                normalizeMangoQuery(
+                    collection.internalStorageInstance.schema,
+                    {
+                        selector: {
+                            firstName: 'a',
+                            lastName: 'b'
+                        }
+                    }
+                )
+            );
+            assert.deepStrictEqual(preparedQuery.queryPlan.index, ['firstName', 'lastName']);
+            collection.database.destroy();
+        });
+    });
+    describe('special case', () => {
+        it('server must be able to iterate with additional fields', async () => {
+            const myIdx = ['firstName', '_meta.lwt', 'passportId'];
+            const collection = await createCollectionWithInternalIndexes([myIdx], 10);
+            const writeResult = await collection.bulkInsert(new Array(10).fill(0).map(() => schemaObjects.human(undefined, undefined, 'alice')));
+            assert.deepStrictEqual(writeResult.error, []);
+            const query = getChangedDocumentsSinceQuery(
+                collection.storageInstance,
+                2
+            );
+            query.selector.firstName = { $eq: 'alice' };
+            const preparedQuery = prepareQuery(
+                collection.storageInstance.schema,
+                query
+            );
+            console.dir(preparedQuery);
+            assert.deepStrictEqual(preparedQuery.queryPlan.index, [
+                'firstName',
+                '_meta.lwt',
+                'passportId'
+            ]);
+
+            const result = await collection.storageInstance.query(preparedQuery);
+            result.documents.forEach(d => assert.strictEqual(d.firstName, 'alice'));
+
+            collection.database.destroy();
+        });
+    });
+});